requirements.txt
Pillow>=5.3.0
beautifulsoup4>=4.6.3
requests>=2.21.0